Drawing, Player Shit
Player ship is now drawn, testing area is being worked on. Bullet class
being worked on, next will be enemies, and then hitboxes?

+++ Living.java
@@ -1,6 +1,6 @@
 public class Living {
   public int health = 1;
-  public int position[] = {250,250};
+  public float position[] = {250,250};
   public int sprite = 0;
   public float rotation = 0;
   public Living() {
